Move rewrite to conventional location.
`scalafix scalacenter/sbtfix/v1` expands into `fix/Scalafix_v1.scala`.
It's possible to keep the original location if users write the full url
of the raw rewrite, `scalafix https://raw.githubusercontent.com/...`,
but I'd like to try and use this convention everywhere.
